==> step2_feature_create.ipynb <==
import os
import pandas as pd
import numpy as np
from glob import glob
from tqdm import tqdm
tqdm.pandas()

import pickle
import pyarrow.parquet as pq

import ipyparallel
# ipcluster nbextension enable --user

DATA = './data'
DATA_OWN = './data_own'
CLICKSTREAM = 'alfabattle2_abattle_clickstream'

Если в test поставить последнюю  по времени, транзакцию из train, то получим 0.36081 на LB

Как вариант, эти таймштуки можно различными преобразованиями (фурье и т.д.) превратить в картинки, а дальше как на работе))

hour_diff

типа hour_diff_mean одна из сильных фичей

Есть файлик, в имени которого есть "train". В нём для каждого клиента есть набор сессий (1) с multi_class_target.

Есть файлики в формате parquet, в них для каждого клиента есть набор сессий (2).

Набор (2) содержит в себе весь набор (1), но набор (2) почти в два раза больше набора (1)


1) left join (2) по идее должно решать эту проблему

FTRS = ['last_target'# last known target for this user (by last session)
        ,'last_moment_invest'
        ,'last_moment_mobile_recharge'
        ,'last_moment_statement'
        ,'last_moment_chat'
        ,'last_moment_card2card_transfer'
        ,'last_moment_phone_money_transfer'
        ,'last_moment_card_recharge'
        ,'last_moment_main_screen'
        ,'last_moment_own_transfer'
        ,'last_moment_credit_info'
        
         #most frequent in this 
        ,'mf_dow'    # day of week
        ,'mf_dom'    # day of month
        ,'mf_pod'    # part of day
        
        #most frequent for this user in this 
         'mf_this_user_dow'    # day of week
        ,'mf_this_user_dom'    # day of month
        ,'mf_this_user_pod'    # part of day
        
        # days from last goal was a 
        ,'days_since_last_invest'        # invest
        ,'days_since_last_statement'     # statement
        ,'days_since_last_chat'          # chat
        ,'days_since_last_card_recharge' # card_recharge
        ,'days_since_last_main_screen'   # main_screen
        ,'days_since_last_own_transfer'  # own_transfer
        ,'days_since_last_credit_info'   # credit_info
        ,'days_since_last_mobile_recharge'      # mobile_recharge
        ,'days_since_last_card2card_transfer'   # card2card_transfer
        ,'days_since_last_phone_money_transfer' # phone_money_transfer
    
        
        #, last several days
       ]

invest                  8055
mobile_recharge         8028
statement               7995
chat                    7989
card2card_transfer      7965
phone_money_transfer    7941
card_recharge           7936
main_screen             7893
own_transfer            7759
credit_info             7707

possible_targets = ['invest', 'mobile_recharge', 'statement', 'chat', 'card2card_transfer', 
                    'phone_money_transfer', 'card_recharge', 'main_screen', 'own_transfer', 
                    'credit_info']

files = sorted(glob(os.path.join(DATA, CLICKSTREAM, 'part*')))
#files

sample = pd.read_csv(os.path.join(DATA, 'alfabattle2_abattle_sample_prediction.csv'))
test   = pd.read_csv(os.path.join(DATA, 'alfabattle2_prediction_session_timestamp.csv'))

targets_df = pd.read_csv(os.path.join(DATA, 'alfabattle2_abattle_train_target.csv'), parse_dates=['timestamp'])
targets_df.sort_values(by='timestamp', ascending = True, inplace = True)

# session wich has a target
sess_with_target = list(targets_df.session_id.values)
uniq_client_pin = list(targets_df.client_pin.unique())

summarize_df = pd.DataFrame(index = range(len(uniq_client_pin)), columns = FTRS)

summarize_df['client_pin'] = uniq_client_pin

data = pd.read_parquet(files[0])
data.sort_values(by='timestamp', ascending=True, inplace = True)
data.shape

len(data.client.unique())

timestamp = pd.Timestamp('2000-01-01')

for client in tqdm(data.client.unique()):
    index = int(summarize_df[summarize_df.client_pin == client].index[0])
    #print(index)
    for targ in possible_targets:
        tmp_list = list(targets_df.query('client_pin == @client & multi_class_target == @targ').timestamp)
        if len(tmp_list) == 0:
            hrs_mean = 0.0
            hrs_std = 0.0
            last = timestamp
        else:
            diff_list = [0]*(len(tmp_list) - 1)
            for idx in range(1, len(tmp_list)):
                diff_list[idx - 1] = (tmp_list[idx] - tmp_list[idx-1]).total_seconds() // 3600

            hrs_mean = np.mean(diff_list)
            hrs_std = np.std(diff_list)
            last = tmp_list[-1]
            
        summarize_df.loc[index, 'hrs_between_' + targ + '_mean'] = hrs_mean
        summarize_df.loc[index, 'hrs_between_' + targ + '_std'] = hrs_std
        summarize_df.loc[index, 'last_moment_' + targ] = last
        
    
            
print('done')    

summarize_df[summarize_df.client_pin == data.client.unique()[0]]

# calculate one in second

%%time
# collect statistic about every client
cnt = 0
everyday = []
for client in tqdm(data.client.unique()[:50]):
    # for each session
    current = ''
    weekend_start_hour = []
    weekday_start_hour = []
    everyday_cnt = []
    for idx in data[data.client == client].index:
        if  (data.loc[idx, 'session_id'] != current):# and (data.loc[idx, 'session_id'] in sess_with_target):
            cnt+= 1
            current = data.loc[idx, 'session_id']
            if data.loc[idx, 'timestamp'].weekday() >= 5:
                weekend_start_hour.append(data.loc[idx, 'timestamp'].hour)
            else:
                weekday_start_hour.append(data.loc[idx, 'timestamp'].hour)
            
            everyday_cnt.append(data.loc[idx, 'timestamp'].weekday())
            
    everyday.append(everyday_cnt)

print(cnt)

len(data.session_id.unique())

pred_sess = ''
dow = []
dom = [] 
weekend_start_hour = []
weekday_start_hour = []

#for idx, el in enumerate(data.session_id):
#for idx in tqdm(range(data.shape[0])):
for idx in tqdm(range(5000)):
    #if (el != pred_sess) and (el in sess_with_target):
    if ( data.loc[idx, 'session_id'] != pred_sess) and (data.loc[idx, 'session_id'] in sess_with_target):
        pred_sess = data.loc[idx, 'session_id']
        if (data.loc[idx, 'timestamp'].weekday() >= 5):
            weekend_start_hour    
        #get start time
        #start_hour.append(data.loc[idx, 'timestamp'].hour)
        #get start day of week
        #dow.append(data.loc[idx, 'timestamp'].weekday())
        #get start day of month
        #dom.append(data.loc[idx, 'timestamp'].day)
        #print(el)
        break

data.loc[0, 'timestamp'].weekday()

data_target[data_target.session_id == data.loc[0, 'session_id']]

data_target.keys()

data.loc[1920, 'session_id'] in sess_with_target

data.loc[1920, 'session_id']

%%time
data_target['dow'] = data_target.timestamp.map(lambda x: x.weekday())
data_target['dom'] = data_target.timestamp.map(lambda x: x.day)
data_target['start_hour'] = data_target.timestamp.map(lambda x: x.hour)

data_target.head()

data_target.groupby('dow').multi_class_target.agg(pd.Series.mode)

realy = data_target.groupby('dow').multi_class_target.value_counts()

realy

data_target.groupby('dom').multi_class_target.agg(pd.Series.mode)

data_target.groupby('start_hour').multi_class_target.value_counts()

data_target.multi_class_target.value_counts().index[1]
